make price decimal instead of integer; fix spelling of description in item model

+++ database/migrations/2021_05_17_105102_create_items_table.php
@@ -17,7 +17,7 @@ class CreateItemsTable extends Migration
             $table->id();
             $table->string('name');
             $table->string('description')->nullable();
-            $table->integer('price');
+            $table->decimal('price',5,2);
             $table->boolean('isOwned');
             $table->string('status');
             $table->date('date_of_purchase');
